+++ app-router/app/meals/share/page.js
@@ -1,24 +1,10 @@
+'use client';
+
 import ImagePicker from '@/components/meals/image-picker';
 import classes from './page.module.css';
+import { shareMeal } from '@/lib/action';
 
 export default function ShareMealPage() {
-    async function shareMeal(formData) {
-        'use server';
-        console.log('done');
-
-        const meal = {
-            title: formData.get('title'),
-            summary: formData.get('summary'),
-            instructions: formData.get('instructions'),
-            image: formData.get('image'),
-            creator: formData.get('name'),
-            creator_email: formData.get('email')
-        }
-
-        console.log(meal);
-
-    }
-
 
     return (
         <>
